Seperate block chain code to a different file and near completed p2p connection
